Python/dpr_python_nvidia/requirements.txt
numpy
scipy
pillow
tifffile
imageio
matplotlib
cupy